+++ ark/docs/content/docs/configuration/index.mdx
@@ -425,7 +425,7 @@ const data = type.string("foo")
 const bad = type.string(5)
 ```
 
-### custom
+### metadata
 
 Additional arbitrary metadata can also be associated with a Type.
