+++ ark/docs/content/docs/configuration/meta.json
@@ -5,6 +5,9 @@
 		"[errors](/docs/configuration#errors)",
 		"[clone](/docs/configuration#clone)",
 		"[onUndeclaredKey](/docs/configuration#onundeclaredkey)",
-		"[jitless](/docs/configuration#jitless)"
+		"[jitless](/docs/configuration#jitless)",
+		"[onFail](/docs/configuration#onFail)",
+		"[prototypes](/docs/configuration#protoypes)",
+		"[metadata](/docs/configuration#metadata)"
 	]
 }
